+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.3]
+
+## Added
+
+- Cancel Generation Button
+
+## Fixed
+
+- Check error logs coming from Ollama and send it to the frontend
+
 ## [1.0.2]
 
 ## Added